Create Training Image
This is the training image for the suggested Universal Wire Standard. I've used Yellow for simplicity's sake in the example, but in practice I believe making the pattern reflective or a color that contrasts well with the individual wire is best. Reflective would improve visibility in low lighting while raising the patter on the wire would allow for haptic confirmation of wire identity.